+++ test/com/cloudogu/ces/cesbuildlib/DoguRegistryTest.groovy
@@ -1,0 +1,13 @@
+package com.cloudogu.ces.cesbuildlib
+
+class DoguRegistryTest extends GroovyTestCase {
+
+    void testCreateRegistryObjectWithDefaults() {
+        // given
+        // when
+        DoguRegistry sut = new DoguRegistry("script")
+
+        // then
+        assertTrue(sut != null)
+    }
+}